requirements.txt
Package               Version
--------------------- ---------
certifi               2022.12.7
charset-normalizer    3.0.1
colorama              0.4.6
decorator             4.4.2
idna                  3.4
imageio               2.25.0
imageio-ffmpeg        0.4.8
moviepy               1.0.3
numpy                 1.23.5
opencv-contrib-python 4.7.0.68
Pillow                9.4.0
pip                   22.3.1
proglog               0.1.10
PyAudio               0.2.13
requests              2.28.2
setuptools            58.1.0
tqdm                  4.64.1
urllib3               1.26.14
